+++ src/main.rs
@@ -6,7 +6,7 @@ use std::{
     time::{Duration, Instant},
 };
 
-use anyhow::Context;
+use anyhow::{bail, Context};
 use clap::Parser;
 use ignore::WalkBuilder;
 use notify::{event::ModifyKind, Event, EventKind, FsEventWatcher, RecursiveMode, Watcher};
@@ -20,17 +20,10 @@ use tracing::{debug, error, warn};
 /// an &[`Event`] and returns a [`bool`] which indicates whether the `event`
 /// should trigger running the command.
 fn should_event_trigger(event: &Event) -> bool {
-    match event.kind {
-        // Drop metadata change events
-        EventKind::Modify(ModifyKind::Metadata(_)) => false,
-        // Drop access events
-        EventKind::Access(_) => false,
-        EventKind::Modify(_) => true,
-        EventKind::Create(_) => true,
-        EventKind::Remove(_) => true,
-        EventKind::Other => true,
-        EventKind::Any => true,
-    }
+    !matches!(
+        event.kind,
+        EventKind::Modify(ModifyKind::Metadata(_)) | EventKind::Access(_)
+    )
 }
 
 /// Run the specified command + args and log any errors that occur.
@@ -43,7 +36,7 @@ async fn run_command(command: &OsStr, args: &[OsString]) {
             None => error!("command exited without code"),
         },
         Err(e) => {
-            error!("{:#}", e)
+            error!("{:#}", e);
         }
     }
 }
@@ -81,13 +74,13 @@ impl DebounceTimer {
 
     /// Stops the timer.
     fn stop(&mut self) {
-        self.start = None
+        self.start = None;
     }
 
     /// Starts the timer if it wasn't previously started.
     fn start_if_stopped(&mut self) {
         if self.start.is_none() {
-            self.start = Some(Instant::now())
+            self.start = Some(Instant::now());
         }
     }
 }
@@ -100,7 +93,7 @@ fn watch_new_files(watcher: &mut FsEventWatcher, event: &Event) {
             if path.exists() {
                 debug!("watching '{}'", path.display());
                 if let Err(e) = watcher.watch(path, RecursiveMode::NonRecursive) {
-                    error!("Failed to watch {}: {:#}", path.display(), e)
+                    error!("Failed to watch {}: {:#}", path.display(), e);
                 }
             }
         }
@@ -113,7 +106,7 @@ fn watch_new_files(watcher: &mut FsEventWatcher, event: &Event) {
 /// event is the configurable debounce time.
 #[derive(Parser)]
 #[clap(trailing_var_arg = true)]
-struct Args {
+struct Opts {
     /// Debounce time in milliseconds
     #[clap(short, long = "debounce", default_value_t = 50)]
     debounce_ms: u64,
@@ -123,34 +116,40 @@ struct Args {
     /// Paths to watch
     #[clap(short, long, default_value = ".")]
     path: Vec<PathBuf>,
-    /// Command to run
-    command: OsString,
-    /// Args for command
-    args: Vec<OsString>,
+    /// Command to run with any arguments
+    #[clap(required = true)]
+    command_to_run: Vec<OsString>,
 }
 
 #[tokio::main]
-async fn try_main(args: Args) -> anyhow::Result<()> {
+async fn try_main(opts: Opts) -> anyhow::Result<()> {
     let (snd, mut rcv) = tokio::sync::mpsc::unbounded_channel();
     // Initialize fs watcher
     let mut watcher = notify::recommended_watcher(move |res| match res {
         Ok(event) => {
             if should_event_trigger(&event) {
                 if let Err(e) = snd.send(event) {
-                    error!("send error: {:?}", e)
+                    error!("send error: {:?}", e);
                 }
             }
         }
         Err(e) => error!("watch error: {:?}", e),
     })?;
 
+    if opts.command_to_run.is_empty() {
+        bail!("No command given")
+    }
+    let mut command_to_run = opts.command_to_run;
+    let command = command_to_run.remove(0);
+    let args = command_to_run;
+
     // Use `ignore` to walk all given paths to add watch events.
-    let mut path_iter = args.path.iter();
+    let mut path_iter = opts.path.iter();
     let mut walk_builder = WalkBuilder::new(path_iter.next().expect("at least one path required"));
     for path in path_iter {
         walk_builder.add(path);
     }
-    if args.skip_ignore_files {
+    if opts.skip_ignore_files {
         walk_builder
             .ignore(false)
             .git_ignore(false)
@@ -163,7 +162,7 @@ async fn try_main(args: Args) -> anyhow::Result<()> {
         watcher.watch(entry.path(), RecursiveMode::NonRecursive)?;
     }
 
-    let mut debounce = DebounceTimer::new(Duration::from_millis(args.debounce_ms));
+    let mut debounce = DebounceTimer::new(Duration::from_millis(opts.debounce_ms));
     loop {
         match debounce.timeout(rcv.recv()).await {
             Ok(Some(event)) => {
@@ -182,18 +181,18 @@ async fn try_main(args: Args) -> anyhow::Result<()> {
                 break;
             }
             Err(_) => {
-                debug!("timeout reached. running command")
+                debug!("timeout reached. running command");
             }
         };
         debounce.stop();
-        run_command(&args.command, &args.args).await;
+        run_command(&command, &args).await;
     }
 
     Ok(())
 }
 
 fn main() {
-    let args = Args::parse();
+    let args = Opts::parse();
     tracing_subscriber::fmt().init();
     if let Err(e) = try_main(args) {
         eprintln!("{:#?}", e);