+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.1.1
+
+- Improve command line argument handling for the specified command. Now `seedo`
+  will stop trying to parse arguments after it sees the command.
+
 ## v0.1.0
 
 - Initial Release